+++ CHANGELOG.md
@@ -10,8 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  * Testing in Travis for python 3.6, 3.7 and 3.8
  * Support in API to timeZone flag for Devo API
  * Support in API for new devo custom formats
- * Support for zip and buffer flags in Sender CLI
- * Functions to change buffer size of Sender
+ * Functions to change buffer size and compression_level of Sender
+ * Support for zip, buffer and compression_level flags in Sender CLI
+
  
 ### Changed
  * SSL Server support to adapt it from python 3.5 to python 3.8